# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the grid platform",
        "title": "PLATFORM API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/createorder": {
            "post": {
                "description": "create an order",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CreateOrder"
                ],
                "summary": "Create order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user address",
                        "name": "userAddress",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "cpAddress",
                        "name": "cpAddress",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "num cpu",
                        "name": "numCPU",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "price cpu",
                        "name": "priCPU",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "num",
                        "name": "numGPU",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "price",
                        "name": "priGPU",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "num",
                        "name": "numStore",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "price",
                        "name": "priStore",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "num",
                        "name": "numMem",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "price",
                        "name": "priMem",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "duration",
                        "name": "duration",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "regist OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/listcp/": {
            "get": {
                "description": "list all providers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Listcps"
                ],
                "summary": "List all providers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/routes.CPInfo"
                            }
                        }
                    },
                    "404": {
                        "description": "page not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/listorder": {
            "get": {
                "description": "list an order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ListOrder"
                ],
                "summary": "List order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user or provider",
                        "name": "role",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "address",
                        "name": "address",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "list OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/listtransfer": {
            "get": {
                "description": "List all transfers of an address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "List transfers"
                ],
                "summary": "List all transfers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "address to show list",
                        "name": "address",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "list transfer OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/pay": {
            "post": {
                "description": "Pay to credit with a transfer's key",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pay"
                ],
                "summary": "Pay for credit",
                "parameters": [
                    {
                        "type": "string",
                        "description": "transfer key",
                        "name": "transkey",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "pay OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/querycredit": {
            "get": {
                "description": "Query credit of someone",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QueryCredit"
                ],
                "summary": "QueryCredit",
                "parameters": [
                    {
                        "type": "string",
                        "description": "role of this caller",
                        "name": "role",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "address of this caller",
                        "name": "address",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "query OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "routes.CPInfo": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "endpoint": {
                    "type": "string"
                },
                "name": {
                    "description": "provider name",
                    "type": "string"
                },
                "numCPU": {
                    "type": "string"
                },
                "numGPU": {
                    "type": "string"
                },
                "numMem": {
                    "type": "string"
                },
                "numStore": {
                    "type": "string"
                },
                "priCPU": {
                    "type": "string"
                },
                "priGPU": {
                    "type": "string"
                },
                "priMem": {
                    "type": "string"
                },
                "priStore": {
                    "type": "string"
                }
            }
        }
    }
}